Update user field in Search model to BigInteger

+++ job_assistant/crawlers/models.py
@@ -62,7 +62,7 @@ class JobAd(SoftDeleteModel):
 
 
 class Search(models.Model):
-    user = models.IntegerField()
+    user = models.BigIntegerField()
     categories = models.CharField(blank=True, null=True)
     technologies = models.CharField(blank=True, null=True)
     workplaces = models.CharField(blank=True, null=True)
